Move all attrs code to attrs_plugin.py
